Handle error for GetIP instead ignoring it
Once we create the VM  we are waiting for 5 seconds to atleast libvirt to kick in
and starting the VM, Also `if s != state.Running`  is mostly to catch other state
of the VM like crashed.

If that 5 sec is not enough for inital libvirt kick in we need to figure out what
could be better.

- https://github.com/code-ready/machine/blob/master/libmachine/state/state.go