Don't unmarshal empty status file.
Currently we are trying to unmarshal the status file which can be empty
if within 5 sec once instance is in running state but not able to update
all the resources here it is network one which located to the libvirt
dir `/var/lib/libvirt/dnsmasq/crc.status`. What we should do is check
if that file is empty and do nothing until the update happen to it.

If we don't ignore the empty file then Unmarshal going to fail with
error and it return the error to GetIP which then make the driver exit.